% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ascii.R
\name{import_ascii}
\alias{import_ascii}
\title{Import ASCII datasets using SPSS syntax files}
\usage{
import_ascii(
  data_file,
  setup_file,
  total_cards = 1,
  var_names,
  var_cards = 1,
  var_positions,
  var_widths,
  card_pattern,
  respondent_pattern
)
}
\arguments{
\item{data_file}{A path to an ASCII data file.}

\item{setup_file}{A path to the corresponding SPSS setup file.}

\item{total_cards}{For multicard files, the number of cards in the file.}

\item{var_names}{A string vector of variable names.}

\item{var_cards}{For multicard files, a numeric vector of the cards on which \code{var_names} are recorded.}

\item{var_positions}{A numeric vector of the column positions in which \code{var_names} are recorded.}

\item{var_widths}{A numeric vector of the widths used to record \code{var_names}.}

\item{card_pattern}{For use when the file does not contain a line for every card for every respondent (or contains extra lines that correspond to no respondent), a regular expression that matches the file's card identifier; e.g., if the card number is stored in the last digit of each line, "\\d$".}

\item{respondent_pattern}{For use when the file does not contain a line for every card for every respondent (or contains extra lines that correspond to no respondent), a regular expression that matches the file's respondent identifier; e.g., if the respondent number is stored in the first four digits of each line, preceded by a space, "(?<=^\\s)\\d{4}".}
}
\value{
A data frame
}
\description{
\code{import_ascii} imports ASCII data files using their associated SPSS setup files.
}
\details{
Many older survey datasets are available only in ASCII format, which is notoriously difficult to work with.  The `import_ascii` function imports uses the SPSS setup file to import data stored in this archaic format.

See \code{\link[readroper]{read_rpr}} and \code{\link[ropercenter]{read_roper}} for implementations that read single, specified variables.
}
\examples{

}
